+++ AZ3166/AZ3166-1.0.0/cores/arduino/drivers/Sensors/RGB_LED.h
@@ -8,7 +8,7 @@ class RGB_LED
     public:
         RGB_LED(PinName red = PB_4, PinName green = PB_3, PinName blue = PC_7);
 
-        void setColor(uint8_t red = 255, uint8_t green = 255, uint8_t blue = 255);
+        void setColor(int red = 255, int green = 255, int blue = 255);
         void turnOff();
 
     private:
